+++ sdk/datafactory/arm-datafactory/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History
     
+## 10.5.0 (2022-04-24)
+    
+**Features**
+
+  - Added Interface PrivateEndpoint
+  - Interface PrivateLinkConnectionApprovalRequest has a new optional parameter privateEndpoint
+    
+    
 ## 10.4.0 (2022-04-20)
     
 **Features**